requirements.txt
streamlit
tensorflow-cpu
requests
numpy
Pillow
pandas